refactor front unused file and code
