+++ frontend/src/headerComponents/NavigationBar.tsx
